+++ app/views/shared/_flash.html.erb
@@ -1,0 +1,6 @@
+<% if notice %>
+  <div class="notice"><%= notice %></div>
+<% end %>
+<% if alert %>
+  <div class="alert"><%= alert %></div>
+<% end %>
